requirements.txt
pandas>=2.0.0
prophet>=1.1.5
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn>=0.30.1
requests>=2.32.3
google-genai>=1.7.0
python-multipart>=0.0.9
plotly>=5.22.0
pillow>=10.4.0
httpx>=0.27.0
python-dotenv>=1.0.1
